refactor: extract trusted renderer IPC into ipc module
Made-with: Cursor
